refactor: move form helpers from item.js to form.js
