+++ src/encoded/static/components/form.js
@@ -5,8 +5,16 @@ var ReactForms = require('react-forms');
 var parseAndLogError = require('./mixins').parseAndLogError;
 var closest = require('../libs/closest');
 var offset = require('../libs/offset');
+var fetched = require('./fetched');
+var inputs = require('./inputs');
+var layout = require('./layout');
 var ga = require('google-analytics');
 var _ = require('underscore');
+
+var Layout = layout.Layout;
+var ItemPreview = inputs.ItemPreview;
+var ObjectPicker = inputs.ObjectPicker;
+var FileInput = inputs.FileInput;
 
 
 var filterValue = function(value) {
@@ -45,6 +53,207 @@ var makeValidationResult = function(validation) {
 };
 
 
+var RepeatingItem = React.createClass({
+
+  render: function() {
+    return (
+      <div {...this.props} className="rf-RepeatingFieldset__item">
+        {this.props.children}
+        <button
+          onClick={this.onRemove}
+          type="button"
+          className="rf-RepeatingFieldset__remove">&times;</button>
+      </div>
+    );
+  },
+
+  onRemove: function(e) {
+    if (!confirm('Are you sure you want to remove this item?')) {
+        e.preventDefault();
+        return;
+    }
+    if (this.props.onRemove) {
+      this.props.onRemove(this.props.name);
+    }
+  }
+
+});
+
+
+var FetchedFieldset = React.createClass({
+
+    getInitialState: function() {
+        var value = this.props.value;
+        var url = typeof value.value == 'string' ? value.value : null;
+        var externalValidation = value.externalValidation;
+        return {
+            url: url,
+            collapsed: url && !externalValidation.isFailure,
+        };
+    },
+
+    render: function() {
+        var ReactForms = require('react-forms');
+        var schema = this.props.schema;
+        var value = this.props.value;
+        var externalValidation = value.externalValidation;
+        var isFailure = externalValidation.isFailure;
+        externalValidation = isFailure ? externalValidation : null;
+        value = value.value;
+        var url = typeof value == 'string' ? value : null;
+        var preview, fieldset;
+
+        if (this.state.url) {
+            var previewUrl = this.state.url;
+            preview = (
+                <fetched.FetchedData>
+                    <fetched.Param name="data" url={previewUrl} />
+                    <ItemPreview />
+                </fetched.FetchedData>
+            );
+            fieldset = (
+                <fetched.FetchedData>
+                    <fetched.Param name="defaultValue" url={this.state.url + '?frame=edit'} />
+                    <ReactForms.Form schema={schema} onUpdate={this.onUpdate}
+                                     externalValidation={externalValidation} />
+                </fetched.FetchedData>
+            );
+        } else {
+            preview = (
+                <ul className="nav result-table">
+                  <li>
+                    <div className="accession">{'New ' + schema.props.get('label')}</div>
+                  </li>
+                </ul>
+            );
+            fieldset = <ReactForms.Form
+                defaultValue={value} schema={schema} onUpdate={this.onUpdate}
+                externalValidation={externalValidation} />;
+        }
+
+        return (
+            <div className="collapsible">
+                <span className="collapsible-trigger" onClick={this.toggleCollapsed}>{this.state.collapsed ? '▶ ' : '▼ '}</span>
+                {isFailure && <ReactForms.Message>{externalValidation.error}</ReactForms.Message>}
+                <div style={{display: this.state.collapsed ? 'block' : 'none'}}>{preview}</div>
+                <div style={{display: this.state.collapsed ? 'none' : 'block'}}>{fieldset}</div>
+            </div>
+        );
+    },
+
+    toggleCollapsed: function() {
+        this.setState({collapsed: !this.state.collapsed});
+    },
+
+    onUpdate: function(value) {
+        value = value.set('@id', this.state.url);
+        this.props.value.setSerialized(value);
+    }
+
+});
+
+
+var jsonSchemaToFormSchema = function(attrs) {
+    var ReactForms = require('react-forms');
+    var schemas = attrs.schemas,
+        p = attrs.jsonNode,
+        props = attrs.props,
+        id = attrs.id,
+        skip = attrs.skip || [];
+    if (props === undefined) {
+        props = {};
+    }
+    if (p.title) props.label = p.title;
+    if (p.description) props.hint = p.description;
+    if (p.type == 'object') {
+        if (p.formInput == 'file') {
+            props.input = <FileInput />;
+            return ReactForms.schema.Scalar(props);
+        } else if (p.formInput == 'layout') {
+            props.input = <Layout editable={true} />;
+            return ReactForms.schema.Scalar(props);
+        } else {
+            props.component = <ReactForms.Fieldset className={props.required ? "required" : ''} />;
+        }
+        var properties = {}, name;
+        for (name in p.properties) {
+            if (name == 'uuid' || name == 'schema_version') continue;
+            if (p.properties[name].calculatedProperty) continue;
+            if (_.contains(skip, name)) continue;
+            var required = _.contains(p.required || [], name);
+            var subprops = {required: required};
+            properties[name] = jsonSchemaToFormSchema({
+                schemas: schemas,
+                jsonNode: p.properties[name],
+                props: subprops,
+            });
+        }
+        return ReactForms.schema.Mapping(props, properties);
+    } else if (p.type == 'array') {
+        props.component = <ReactForms.RepeatingFieldset className={props.required ? "required" : ""} item={RepeatingItem} />;
+        return ReactForms.schema.List(props, jsonSchemaToFormSchema({schemas: schemas, jsonNode: p.items}));
+    } else if (p.type == 'boolean') {
+        props.type = 'bool';
+        return ReactForms.schema.Scalar(props);
+    } else {
+        if (props.required) props.component = <ReactForms.Field className="required" />;
+        if (p.pattern) {
+            props.validate = function(schema, value) { return (typeof value == 'string') ? value.match(p.pattern) : true; };
+        }
+        if (p['enum']) {
+            var options = p['enum'].map(v => <option value={v}>{v}</option>);
+            if (!p.default) {
+                options = [<option value={null} />].concat(options);
+            }
+            props.input = <select className="form-control">{options}</select>;
+        }
+        if (p.linkTo) {
+            var restrictions = {type: [p.linkTo]};
+            props.input = (
+                <ObjectPicker searchBase={"?mode=picker&type=" + p.linkTo} restrictions={restrictions} />
+            );
+        } else if (p.linkFrom) {
+            // Backrefs have a linkFrom property in the form
+            // (object type).(property name)
+            var a = p.linkFrom.split('.'), linkType = a[0], linkProp = a[1];
+            // Get the schema for the child object, omitting the attribute that
+            // refers to the parent.
+            var linkFormSchema = jsonSchemaToFormSchema({
+                schemas: schemas,
+                jsonNode: schemas[linkType],
+                skip: [linkProp]
+            });
+            // Use a special FetchedFieldset component which can take either an IRI
+            // or a full object as its value, and render a sub-form using the child
+            // object schema.
+            var component = <FetchedFieldset schema={linkFormSchema} />;
+            // Default value for new children needs to refer to the parent.
+            var defaultValue = jsonSchemaToDefaultValue(schemas[linkType]);
+            defaultValue[linkProp] = id;
+            return ReactForms.schema.Scalar({component: component, defaultValue: defaultValue});
+        }
+        if (p.type == 'integer' || p.type == 'number') {
+            props.type = 'number';
+        }
+        if (p.formInput == 'textarea') {
+            props.input = <textarea rows="4" />;
+        }
+        return ReactForms.schema.Scalar(props);
+    }
+};
+
+
+var jsonSchemaToDefaultValue = function(schema) {
+    var defaultValue = {};
+    _.each(schema.properties, function(property, name) {
+        if (property['default'] !== undefined) {
+            defaultValue[name] = property['default'];
+        }
+    });
+    return defaultValue;
+};
+
+
 var Form = module.exports.Form = React.createClass({
     contextTypes: {
         adviseUnsavedChanges: React.PropTypes.func,
@@ -67,7 +276,7 @@ var Form = module.exports.Form = React.createClass({
     getDefaultProps: function() {
         return {
             submitLabel: 'Save',
-        }
+        };
     },
 
     getInitialState: function() {
@@ -230,4 +439,26 @@ var Form = module.exports.Form = React.createClass({
             errors: schemaErrors
         });
     }
+});
+
+
+var JSONSchemaForm = module.exports.JSONSchemaForm = React.createClass({
+
+    getInitialState: function() {
+        var type = this.props.type;
+        var schemas = this.props.schemas;
+        return {
+            schema: jsonSchemaToFormSchema({
+                schemas: schemas,
+                jsonNode: schemas[type],
+                id: this.props.id
+            }),
+            value: this.props.context || jsonSchemaToDefaultValue(schemas[type]),
+        };
+    },
+
+    render: function() {
+        return <Form {...this.props} defaultValue={this.state.value} schema={this.state.schema} />;
+    }
+
 });
